+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 - Fixed spelling mistake in the entrance randomizer option.
 - Fixed spelling mistake in Amy's main theme in songs.json.
 - The spaces between the numbers for emblem requirements in the map now scale properly with the resolution.
+- Capsules broken as Sonic no longer count as broken for Tails when playing as Tails.
+    - This was causing issues in-game and you can avoid this problem in Casinopolis by disabling missable capsules.
 
 ## [v1.2.0 pre-release 2] - 2026-01-19